Move batch wrapper stuff into its own file
