+++ batch_job/src/Makefile
@@ -15,6 +15,7 @@ EXTERNAL_LIBRARIES = ../../work_queue/src/libwork_queue.a $(CHIRP_LIB) ../../dtt
 SOURCES = \
 	batch_file.c \
 	batch_task.c \
+	batch_wrapper.c \
 	batch_job.c \
 	batch_job_amazon.c \
 	batch_job_dryrun.c \
