+++ batch_job/src/batch_task.c
@@ -4,23 +4,9 @@ This software is distributed under the GNU General Public License.
 See the file COPYING for details.
 */
 
-#include <assert.h>
-#include <unistd.h>
-#include <stdlib.h>
-#include <string.h>
-#include <errno.h>
-
 #include "batch_task.h"
 #include "stringtools.h"
 #include "xxmalloc.h"
-#include "path.h"
-#include "debug.h"
-
-struct batch_task_wrapper {
-	struct list *pre;
-	struct list *post;
-	struct list *cmd;
-};
 
 /** Creates new batch_task and initializes file lists. */
 struct batch_task *batch_task_create(struct batch_queue *queue)
@@ -145,122 +131,6 @@ void batch_task_set_info(struct batch_task *t, struct batch_job_info *info)
 	t->info->exit_code = info->exit_code;
 	t->info->exit_signal = info->exit_signal;
 	t->info->disk_allocation_exhausted = info->disk_allocation_exhausted;
-}
-
-struct batch_task_wrapper *batch_task_wrapper_create(void) {
-	return xxcalloc(1, sizeof(struct batch_task_wrapper));
-}
-
-void batch_task_wrapper_delete(struct batch_task_wrapper *w) {
-	if (!w) return;
-	list_free(w->pre);
-	list_delete(w->pre);
-	list_free(w->post);
-	list_delete(w->post);
-	list_free(w->cmd);
-	list_delete(w->cmd);
-	free(w);
-}
-
-void batch_task_wrapper_pre(struct batch_task_wrapper *w, const char *cmd) {
-	assert(w);
-	assert(cmd);
-	assert(!w->cmd);
-
-	if (!w->pre) {
-		w->pre = list_create();
-		assert(w->pre);
-	}
-	list_push_tail(w->pre, string_escape_shell(cmd));
-}
-
-void batch_task_wrapper_argv(struct batch_task_wrapper *w, char *const argv[]) {
-	assert(w);
-	assert(argv);
-	assert(!w->cmd);
-
-	w->cmd = list_create();
-	assert(w->cmd);
-	for (unsigned i = 0; argv[i]; i++) {
-		int rc = list_push_tail(w->cmd, string_escape_shell(argv[i]));
-		assert(rc == 1);
-	}
-}
-
-void batch_task_wrapper_cmd(struct batch_task_wrapper *w, char *const argv[]) {
-	assert(w);
-	assert(argv);
-	assert(!w->cmd);
-
-	w->cmd = list_create();
-	assert(w->cmd);
-	for (unsigned i = 0; argv[i]; i++) {
-		int rc = list_push_tail(w->cmd, string_quote_shell(argv[i]));
-		assert(rc == 1);
-	}
-}
-
-void batch_task_wrapper_post(struct batch_task_wrapper *w, const char *cmd) {
-	assert(w);
-	assert(cmd);
-	assert(!w->cmd);
-
-	if (!w->post) {
-		w->post = list_create();
-		assert(w->post);
-	}
-	list_push_tail(w->post, string_escape_shell(cmd));
-}
-
-char *batch_task_wrapper_write(struct batch_task_wrapper *w, struct batch_task *task, const char *prefix) {
-	assert(w);
-	assert(prefix);
-
-	char *name = string_format("%sXXXXXX", prefix);
-	int wrapper_fd = mkstemp(name);
-	if (wrapper_fd == -1)
-			fatal("failed to create wrapper: %s", strerror(errno));
-
-	batch_task_add_input_file(task, name, NULL);
-
-	if (fchmod(wrapper_fd, 0700) == -1)
-			fatal("failed to make wrapper executable: %s", strerror(errno));
-
-	FILE *wrapper = fdopen(wrapper_fd, "w");
-	if (!wrapper)
-			fatal("failed to open wrapper: %s", strerror(errno));
-
-	fprintf(wrapper, "#!/bin/sh\n");
-	fprintf(wrapper, "set -e\n");
-
-	if (w->post) {
-		// function name unlikely to collide with user's stuff
-		fprintf(wrapper, "CLEANUP_76tnb43rr7 () {\n");
-		list_first_item(w->post);
-		for (const char *c; (c = list_next_item(w->post));) {
-			fprintf(wrapper, "eval %s\n", c);
-		}
-		fprintf(wrapper, "}\n");
-		fprintf(wrapper, "trap CLEANUP_76tnb43rr7 EXIT INT TERM\n");
-	}
-
-	if (w->pre) {
-		list_first_item(w->pre);
-		for (const char *c; (c = list_next_item(w->pre));) {
-			fprintf(wrapper, "eval %s\n", c);
-		}
-	}
-
-	if (w->cmd) {
-		list_first_item(w->cmd);
-		for (const char *c; (c = list_next_item(w->cmd));) {
-			fprintf(wrapper, " %s", c);
-		}
-		fprintf(wrapper, "\n");
-	}
-
-	fclose(wrapper);
-	return name;
 }
 
 /* vim: set noexpandtab tabstop=4: */
